fix(audit): apply PR #196 reviewer 5🟡 — int64 / promptSHA system-only / egress gate / spec sync
Reviewer findings (5🟡) on PR #196:

1. Spec §2 typed LatencyMS/EgressBytes int64; impl used int. Switched
   audit.Entry + reasoner.CallInfo + reasoner.CompleteResult.EgressBytes
   to int64 — portability guard against a 32-bit build target.

2. Spec §2 listed six fields; PR shipped seven (added CacheHit). CacheHit
   wires V7 #183 so retro can distinguish cold/warm calls without
   re-reading Usage — paired spec amendment adds it to §2 (now seven
   fields). Spec also calls out PromptSHA-system-only and egress-gate
   semantics inline.

3. PromptSHA was hashing the assembled persona+system blob, breaking the
   spec §4.1/§4.2 audit-replay contract (sha → SystemPrompt registry
   lookup). Now hashes r.SystemPrompt only — PersonaPrefix is a
   per-workspace overlay tracked elsewhere. New
   TestPromptSHA_ExcludesPersonaPrefix asserts the SHA does not match
   the assembled blob.

4. EgressBytes now gated behind LEAH_AUDIT_EGRESS_BYTES=1 (default OFF).
   The SDK doesn't surface the wire payload, so deriving it costs a
   parallel json.Marshal on every LLM call — opt-in keeps that off the
   hot path.

5. Wave-label: PR shipped W92 audit-schema + W93 reasoner-instrumentation
   work; W94 (CostMonth/Breaker) is a separate PR. Title/body retitle
   handled separately.

go build ./… clean; go vet ./… clean; bash scripts/check.sh exits 0
(density warn-only on pre-existing files).

+++ internal/audit/audit.go
@@ -28,14 +28,16 @@ type Entry struct {
 	// are treated as the implicit "default" workspace by downstream readers.
 	Workspace string `json:"workspace,omitempty"`
 
-	// W94 LLM-dim schema extension (S2 §2). All omitempty so legacy rows
-	// stay byte-identical and pre-W94 readers ignore unknown keys.
+	// W92+W93 LLM-dim schema extension (S2 §2). All omitempty so legacy
+	// rows stay byte-identical and pre-W92 readers ignore unknown keys.
+	// LatencyMS / EgressBytes are int64 per spec §2 — guards against a
+	// hypothetical 32-bit build target silently truncating.
 	Model        string `json:"model,omitempty"`
 	PromptSHA    string `json:"prompt_sha,omitempty"`
 	InputTokens  int    `json:"input_tokens,omitempty"`
 	OutputTokens int    `json:"output_tokens,omitempty"`
-	LatencyMS    int    `json:"latency_ms,omitempty"`
-	EgressBytes  int    `json:"egress_bytes,omitempty"`
+	LatencyMS    int64  `json:"latency_ms,omitempty"`
+	EgressBytes  int64  `json:"egress_bytes,omitempty"`
 	CacheHit     bool   `json:"cache_hit,omitempty"`
 }
